[refactoring] remove #pragma once in cpp files, split ScriptFactory.h & ScriptHandler.h
