+++ Shell/Shell.vcxproj
@@ -158,6 +158,8 @@
     <ClInclude Include="LbaRangeVerifier.h" />
     <ClInclude Include="Logger.h" />
     <ClInclude Include="Read.h" />
+    <ClInclude Include="ScriptFactory.h" />
+    <ClInclude Include="ScriptHandler.h" />
     <ClInclude Include="Shell.h" />
     <ClInclude Include="SsdExcutable.h" />
     <ClInclude Include="SsdHelper.h" />
